refactor: 검증 기능들 역할에 따른 분리
