+++ src/main/java/lotto/validator/InputValidator.java
@@ -1,9 +1,7 @@
 package lotto.validator;
 
-import static lotto.constant.ErrorMessage.NOT_DIVISIBLE_MESSAGE;
 import static lotto.constant.ErrorMessage.NOT_NUMBER_MESSAGE;
-import static lotto.constant.ErrorMessage.OVER_RANGE_MESSAGE;
-import static lotto.constant.LottoConstant.LOTTO_PRICE;
+import static lotto.constant.ErrorMessage.OVER_NUMBER_RANGE_MESSAGE;
 
 import lotto.view.Output;
 
@@ -11,6 +9,7 @@ public class InputValidator {
     //상수들 역할에 따라 분리
 
     private static final String NUMBER_REGEX = "[0-9]+";
+
 
     public static void onlyNumber(String input) {
         if (isNotMatchRegex(input, NUMBER_REGEX)) {
@@ -28,28 +27,22 @@ public class InputValidator {
         return true;
     }
 
-    public static void inRange(String input) {
+    public static void inNumberRange(String input) {
         try {
             Integer.parseInt(input);
         } catch (NumberFormatException e) {
-            Output.errorMessage(OVER_RANGE_MESSAGE);
-            throw new IllegalArgumentException(OVER_RANGE_MESSAGE);
+            Output.errorMessage(OVER_NUMBER_RANGE_MESSAGE);
+            throw new IllegalArgumentException(OVER_NUMBER_RANGE_MESSAGE);
         }
     }
 
-    public static void divisibleUnit(int price) {
-        if (isNotDivisibleUnit(price)) {
-            Output.errorMessage(NOT_DIVISIBLE_MESSAGE);
-            throw new IllegalArgumentException(NOT_DIVISIBLE_MESSAGE);
-        }
-    }
-
-    private static boolean isNotDivisibleUnit(int price) {
-        //상수화를 시키는 것이 오히려 가독성을 떨어뜨리는 느낌?
-        if ((price % LOTTO_PRICE == 0) && (price != 0)) {
-            return false;
-        }
-        return true;
-    }
+//    public static void dividedCountCorrect(String input, String separationSign) {
+//        int dividedCount = Arrays.stream(input.split(separationSign))
+//                .toList()
+//                .size();
+//        if (dividedCount != LottoConstant.LOTTO_NUMBER_COUNT) {
+//            throw new IllegalArgumentException(NOT_MATCH_COUNT_MESSAGE);
+//        }
+//    }
 
 }
